grading-app/requirements.txt
streamlit
google-generativeai==0.3.2
python-dotenv
pdf2image
